Refactored gridElement into different files
