+++ screens/GridScreen.tsx
@@ -2,26 +2,24 @@ import React, { useState } from 'react';
 import {
     StyleSheet,
     Switch,
-
-
     Text,
     View,
 } from 'react-native';
 
-
+import { Icon } from '@rneui/themed';
 
 
 import GridElement from '../components/GridElement';
 
 import { MIDI_HTTP_Service } from '../services/MIDI_HTTP_Service';
 import { ColorPresetService } from '../services/ColorPresetService';
-import { DEFAULT_TEXT_COLOR, DEFAULT_UNPRESSED_COLOR, DEFAULT_PRESSED_COLOR } from '../constants/Colors';
+import { DEFAULT_COLOR_PRESET } from '../constants/Colors';
 
 
 const midiService = new MIDI_HTTP_Service("192.168.0.12", "4848");
 const colorPresetService = new ColorPresetService();
 
-colorPresetService.createColorPreset("Default", DEFAULT_TEXT_COLOR, DEFAULT_UNPRESSED_COLOR, DEFAULT_PRESSED_COLOR);
+colorPresetService.createColorPreset("Default", DEFAULT_COLOR_PRESET.textColor, DEFAULT_COLOR_PRESET.unpressedColor, DEFAULT_COLOR_PRESET.pressedColor);
 colorPresetService.createColorPreset("Alt", '#26ffcc', '#330c29', '#c5ed12');
 
 
@@ -38,8 +36,15 @@ export default function GridScreen() {
             <View style={styles.headerOptions}>
                 <Switch onChange={() => { setIsPlayMode(!isPlayMode) }} value={isPlayMode}></Switch>
                 <Text style={styles.playOrEditText}>{isPlayMode ? "PLAY" : "EDIT"}</Text>
+                {!!!isPlayMode &&
+                    <View style={{}}>
+                        <Icon name='settings' color='#ffffff' style={{}} />
+                    </View>
+                }
             </View>
-            <View style={styles.gridArea}>
+
+
+            <View style={styles.gridArea}>{/* TODO: Make this whole area dynamically genereated based on the grid settings */}
                 <View style={styles.gridElementRow} >
                     <GridElement MIDI_HTTP_Service={midiService} colorPresetService={colorPresetService} initialName="C Note" initialNoteNumber={0} initialOctave={6} isPlayMode={isPlayMode}></GridElement>
                     <GridElement MIDI_HTTP_Service={midiService} colorPresetService={colorPresetService} initialName="Db Note" initialNoteNumber={1} initialOctave={6} isPlayMode={isPlayMode}></GridElement>
